backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
dnspython==2.7.0
python-whois==0.9.4
cryptography==44.0.0
slowapi==0.1.9
